fix: 修复搜索统计租户隔离问题
- 修复 PostgreSQL COUNT 返回 bigint 导致的 InvalidCastException
- 修复 SQL 中 TenantId 模糊引用问题
- 简化搜索统计查询，移除复杂的 JOIN
- 实现正确的租户隔离：租户用户只看自己数据，Host 管理员需指定租户
- 将 ActiveUsers/TodayActiveUsers 类型从 int 改为 long

+++ src/KnowledgeHub.Application.Contracts/Search/Dtos/SearchStatisticsDtos.cs
@@ -18,8 +18,8 @@ public class SearchStatsBreakdown
 {
     public long TotalSearches { get; set; }
     public long TodaySearches { get; set; }
-    public int ActiveUsers { get; set; }
-    public int TodayActiveUsers { get; set; }
+    public long ActiveUsers { get; set; }
+    public long TodayActiveUsers { get; set; }
 }
 
 public class SearchStatsQueryDto
